refactor(orchestrator): extract feedback builder and inject retry budget
Moves buildSupervisorFeedback + sanitizeFeedbackText into their own
module (src/application/orchestrator/feedback.ts) so the formatter can
be unit-tested without pulling in the entire orchestrator graph and its
infra dependencies.

Adds an optional maxRetries parameter to runExecutorWithEval and
runAideWithEval (defaults to EVAL_RETRIES) so tests can exercise every
branch of the loop without fiddling with process.env. Both helpers are
now exported with @internal annotation — consumed by tests, not meant
for external callers.
